Introduce `processhtml.ts`, reorganize modules
